+++ tests/Decompositions/LUTest.php
@@ -25,31 +25,31 @@ class LUTest extends TestCase
      */
     public function decompose3x3() : void
     {
-        $a = Matrix::quick([
+        $a = Matrix::fromArray([
             [22.0, -17.0, 12.0],
             [4.0, 11.0, -2.0],
             [20.0, -6.0, -9.0],
-        ]);
+        ], false);
 
         $lu = LU::decompose($a);
 
-        $l = Matrix::quick([
+        $l = Matrix::fromArray([
             [1.0, 0, 0],
             [0.18181818181818182, 1.0, 0],
             [0.9090909090909091, 0.6709677419354838, 1.0],
-        ]);
+        ], false);
 
-        $u = Matrix::quick([
+        $u = Matrix::fromArray([
             [22.0, -17.0, 12.0],
             [0.0, 14.09090909090909, -4.181818181818182],
             [0.0, 0.0, -17.10322580645161],
-        ]);
+        ], false);
 
-        $p = Matrix::quick([
+        $p = Matrix::fromArray([
             [1.0, 0, 0],
             [0.0, 1.0, 0],
             [0.0, 0.0, 1.0],
-        ]);
+        ], false);
 
         $expected = new LU($l, $u, $p);
 
@@ -67,12 +67,12 @@ class LUTest extends TestCase
      */
     public function decomposeRequiresPivoting() : void
     {
-        $a = Matrix::quick([
+        $a = Matrix::fromArray([
             [0.0,  1.0,  0.0,  0.0],
             [-1.0, 0.0,  0.0,  0.0],
             [0.0,  0.0,  0.0,  2.0],
             [0.0,  0.0,  3.0,  1.0],
-        ]);
+        ], false);
 
         $lu = LU::decompose($a);
 
@@ -88,13 +88,13 @@ class LUTest extends TestCase
      */
     public function decompose1x1() : void
     {
-        $a = Matrix::quick([[9.0]]);
+        $a = Matrix::fromArray([[9.0]], false);
 
         $lu = LU::decompose($a);
 
-        $l = Matrix::quick([[1.0]]);
-        $u = Matrix::quick([[9.0]]);
-        $p = Matrix::quick([[1.0]]);
+        $l = Matrix::fromArray([[1.0]], false);
+        $u = Matrix::fromArray([[9.0]], false);
+        $p = Matrix::fromArray([[1.0]], false);
 
         $expected = new LU($l, $u, $p);
 
@@ -106,27 +106,27 @@ class LUTest extends TestCase
      */
     public function decomposeDiagonal() : void
     {
-        $a = Matrix::quick([
+        $a = Matrix::fromArray([
             [2.0, 0.0],
             [0.0, 3.0],
-        ]);
+        ], false);
 
         $lu = LU::decompose($a);
 
-        $l = Matrix::quick([
+        $l = Matrix::fromArray([
             [1.0, 0.0],
             [0.0, 1.0],
-        ]);
+        ], false);
 
-        $u = Matrix::quick([
+        $u = Matrix::fromArray([
             [2.0, 0.0],
             [0.0, 3.0],
-        ]);
+        ], false);
 
-        $p = Matrix::quick([
+        $p = Matrix::fromArray([
             [1.0, 0.0],
             [0.0, 1.0],
-        ]);
+        ], false);
 
         $expected = new LU($l, $u, $p);
 
@@ -140,10 +140,10 @@ class LUTest extends TestCase
     {
         $this->expectException(InvalidArgumentException::class);
 
-        LU::decompose(Matrix::quick([
+        LU::decompose(Matrix::fromArray([
             [1.0, 2.0, 3.0],
             [4.0, 5.0, 6.0],
-        ]));
+        ], false));
     }
 
     /**
@@ -153,10 +153,10 @@ class LUTest extends TestCase
     {
         $this->expectException(RuntimeException::class);
 
-        LU::decompose(Matrix::quick([
+        LU::decompose(Matrix::fromArray([
             [1.0, 2.0],
             [2.0, 4.0],
-        ]));
+        ], false));
     }
 
     /**
@@ -164,20 +164,20 @@ class LUTest extends TestCase
      */
     public function accessorsReturnMatrices() : void
     {
-        $l = Matrix::quick([
+        $l = Matrix::fromArray([
             [1.0, 0.0],
             [0.5, 1.0],
-        ]);
+        ], false);
 
-        $u = Matrix::quick([
+        $u = Matrix::fromArray([
             [2.0, 3.0],
             [0.0, 4.0],
-        ]);
+        ], false);
 
-        $p = Matrix::quick([
+        $p = Matrix::fromArray([
             [0.0, 1.0],
             [1.0, 0.0],
-        ]);
+        ], false);
 
         $lu = new LU($l, $u, $p);